APIS/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
mangum==0.17.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
boto3==1.34.34
pyjwt==2.8.0
cryptography==42.0.0
python-dotenv==1.0.0
requests==2.31.0
email-validator==2.1.0
